requirements.txt
# rq and statsd/Graphite is used by the job handler
rq==0.12.0
statsd==3.3.0

# yaml is used by file_utils used by ResourceContainer
pyyaml==3.13

# dateutil is used by file_utils used by ResourceContainer
py-dateutil==2.2.0

# sqlalchemy is used by manifest
sqlalchemy==1.2.9
pymysql==0.8.1
# NOTE: pymysql 0.9.2 wouldn't work in docker build

# boto3 is used by aws_tools
boto3==1.7.55
